Remove register_default_gcs() override
As it doesn't work as expected anyway, not magically dragging in symgc
into the build - the object file needs to be linked for the strong
symbol to override druntime's weak default one. And linking the object
file is enough to register the GC anyway, due to the CRT constructor.

Additionally, the strong symbol means that an app cannot define its own,
or link another alternative GC which also provides such a strong symbol.
This e.g. causes a druntime integration test to fail (which overrides
`register_default_gcs()` too, and the linker then errors out with 2
conflicting strong symbols), when integrating symgc into druntime.